Assets/Scripts/pythonScript/requirements.txt
aniso8601==8.0.0
assimp==4.1.4
click==8.1.3
fake-bpy-module-latest==20230804
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.3
gradio_client==0.2.8
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyassimp==5.2.5
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
torch==2.0.1
torchvision==0.15.2
trimesh==3.22.2
Werkzeug==2.3.6
ipywidgets